+++ chaoscf/actions.py
@@ -2,7 +2,6 @@
 import random
 from typing import Any, Dict, List
 
-from chaoslib.exceptions import FailedActivity
 from chaoslib.types import Configuration, Secrets
 from logzero import logger
 
@@ -11,7 +10,7 @@ from chaoscf.api import call_api, get_app_by_name, get_app_routes_by_host, \
 
 __all__ = ["delete_app", "remove_routes_from_app", "terminate_app_instance",
            "terminate_some_random_instance", "unbind_service_from_app",
-           "unmap_route_from_app", "map_route_to_app"]
+           "unmap_route_from_app", "map_route_to_app", "stop_app"]
 
 
 def delete_app(app_name: str, configuration: Configuration, secrets: Secrets,
@@ -27,6 +26,22 @@ def delete_app(app_name: str, configuration: Configuration, secrets: Secrets,
 
     path = "/v2/apps/{a}".format(a=app['metadata']['guid'])
     call_api(path, configuration, secrets, method="DELETE")
+
+
+def stop_app(app_name: str, configuration: Configuration, secrets: Secrets,
+             org_name: str = None, space_name: str = None):
+    """
+    Stop application
+
+    See https://apidocs.cloudfoundry.org/280/apps/updating_an_app.html
+    """
+    app = get_app_by_name(
+        app_name, configuration, secrets, org_name=org_name,
+        space_name=space_name)
+
+    path = "/v2/apps/{a}".format(a=app['metadata']['guid'])
+    call_api(
+        path, configuration, secrets, method="PUT", body={"state": "STOPPED"})
 
 
 def remove_routes_from_app(app_name: str, route_host: str,